third_party/blink/web_tests/fast/css/hover-active-drag-expected.txt

Dragging out of an element should cause it to lose :hover

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS background is "rgb(255, 0, 0)"
PASS background is "rgb(0, 0, 0)"
PASS background is "rgb(255, 255, 0)"
PASS background is "rgb(0, 0, 0)"
PASS background is "rgb(0, 255, 0)"
PASS background is "rgb(255, 0, 0)"
PASS background is "rgb(0, 255, 0)"
PASS background is "rgb(0, 0, 0)"
PASS background is "rgb(0, 0, 0)"
PASS background is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
